+++ examples/optimizing-loaders/src/pages/index.js
@@ -1,3 +1,4 @@
+/* globals DEFINED */
 import React from 'react';
 import _ from 'lodash';
 export default class Home extends React.PureComponent {
@@ -14,11 +15,13 @@ export default class Home extends React.PureComponent {
     return (
       <div>
         <h1>Optimization loaders and plugins</h1>
-
+        Did the DefinePlugin work?{' '}
+        <span style={{ fontWeight: 'bold' }}>
+          {DEFINED === 'yes' ? 'Yes!' : 'No!'}
+        </span>
         <div style={{ marginTop: 30 }}>
           <img src={require('../img/man-in-a-bottle.jpg')} />
         </div>
-
         <div style={{ marginTop: 30 }}>
           The Lodash plugin works if the following does not say "3":{' '}
           <code style={{ background: '#000', color: '#fff', padding: '1em' }}>