+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Feature:** Added a helpful message if the provided port is not available.
 - **Feature:** Files reference as URLs in CSS then copied into the output directory (via postcss-url) now include the file's original basename as well as the hash.
+- **Feature:** Expose Batfish's version of Webpack on `require('@mapbox/batfish').webpack`, so it can be used for custom plugin configuration.
 - **Fix:** Prevent batfish from entering an unpredictable state if provided port is unavailable.
 
 ## 1.8.2